reverted the nightly publishing changes.
discovered that the android (phonegap) publishing is not done by the duckietv-builder-publish js as I had a right to expect, but is in fact done by the duckietv-builder-make-binaries js. Bah!
which means all by checking code i wrote for the publish js only stops the nighty github release publishing.
doh!